🐛 fix v-ime-guard test coverage gap in mountWithApp
The plugin file default-exported defineNuxtPlugin(...), so importing
its named export for tests also executed that call at module load time
(unhandled side effect outside Nuxt runtime) — this once caused an
unrelated spec (BookmarksLayout) to leak a real network request when a
different mount helper picked up the same module. Split the pure
directive object into app/directives/imeGuard.ts and have the plugin
just import + register it.

mountWithApp now registers 'ime-guard' globally so v-ime-guard-bearing
components render correctly under test. The directive itself only
intercepts Enter via a document-capture-phase listener, which requires
the element to be attached to the real document — so call sites that
actually exercise IME composition (TagsHeader, ChatBot, EditTag) pass
attachTo: document.body. Not defaulting attachTo globally: an earlier
attempt caused unrelated regressions in click-outside/setValue timing
in other specs.
